+++ variables-naming.tf
@@ -1,12 +1,12 @@
 # Generic naming variables
 variable "name_prefix" {
-  description = "Optional prefix for the generated name"
+  description = "Optional prefix for the generated name."
   type        = string
   default     = ""
 }
 
 variable "name_suffix" {
-  description = "Optional suffix for the generated name"
+  description = "Optional suffix for the generated name."
   type        = string
   default     = ""
 }